Trait tokio_timer::clock::Now

pub trait Now: Send + Sync + 'static {
    fn now(&self) -> Instant;
}

Returns Instant values representing the current instant in time.

This allows customizing the source of time which is especially useful for testing.

Implementations must ensure that calls to now return monotonically increasing Instant values.

Required methods

fn now(&self) -> Instant

Returns an instant corresponding to “now”.