Struct tokio_timer::Timeout

pub struct Timeout<T> { /* fields omitted */ }

Allows a Future or Stream to execute for a limited amount of time.

If the future or stream completes before the timeout has expired, then Timeout returns the completed value. Otherwise, Timeout returns an Error.

Futures and Streams

The exact behavor depends on if the inner value is a Future or a Stream. In the case of a Future, Timeout will require the future to complete by a fixed deadline. In the case of a Stream, Timeout will allow each item to take the entire timeout before returning an error.

In order to set an upper bound on the processing of the entire stream, then a timeout should be set on the future that processes the stream. For example:

// import the `timeout` function, usually this is done
// with `use tokio::prelude::*`
use tokio::prelude::FutureExt;
use futures::Stream;
use futures::sync::mpsc;
use std::time::Duration;

let (tx, rx) = mpsc::unbounded();

let process = rx.for_each(|item| {
    // do something with `item`
});

// Wrap the future with a `Timeout` set to expire in 10 milliseconds.
process.timeout(Duration::from_millis(10))

Cancelation

Cancelling a Timeout is done by dropping the value. No additional cleanup or other work is required.

The original future or stream may be obtained by calling Timeout::into_inner. This consumes the Timeout.

Implementations

impl<T> Timeout<T>

pub fn new(value: T, timeout: Duration) -> Timeout<T>

Create a new Timeout that allows value to execute for a duration of at most timeout.

The exact behavior depends on if value is a Future or a Stream.

See type level documentation for more details.

Examples

Create a new Timeout set to expire in 10 milliseconds.

use tokio::timer::Timeout;
use futures::Future;
use futures::sync::oneshot;
use std::time::Duration;

let (tx, rx) = oneshot::channel();

// Wrap the future with a `Timeout` set to expire in 10 milliseconds.
Timeout::new(rx, Duration::from_millis(10))

pub fn get_ref(&self) -> &T

Gets a reference to the underlying value in this timeout.

pub fn get_mut(&mut self) -> &mut T

Gets a mutable reference to the underlying value in this timeout.

pub fn into_inner(self) -> T

Consumes this timeout, returning the underlying value.

impl<T: Future> Timeout<T>

pub fn new_at(future: T, deadline: Instant) -> Timeout<T>

Create a new Timeout that completes when future completes or when deadline is reached.

This function differs from new in that:

It only accepts Future arguments.

It sets an explicit Instant at which the timeout expires.