Module tokio_timer::timer::entry

Structs

Entry

Internal state shared between a Delay instance and the timer.

Time

Stores the info for Delay.

Constants

ELAPSED

Flag indicating a timer entry has elapsed

ERROR

Flag indicating a timer entry has reached an error state

Functions

is_elapsed