Struct tokio_timer::timer::entry::Entry

pub(crate) struct Entry {
    time: CachePadded<UnsafeCell<Time>>,
    inner: Option<Weak<Inner>>,
    state: AtomicU64,
    task: AtomicTask,
    pub(super) queued: AtomicBool,
    pub(super) next_atomic: UnsafeCell<*mut Entry>,
    when: UnsafeCell<Option<u64>>,
    pub(super) next_stack: UnsafeCell<Option<Arc<Entry>>>,
    pub(super) prev_stack: UnsafeCell<*const Entry>,
}

Internal state shared between a Delay instance and the timer.

This struct is used as a node in two intrusive data structures:

An atomic stack used to signal to the timer thread that the entry state has changed. The timer thread will observe the entry on this stack and perform any actions as necessary.

A doubly linked list used only by the timer thread. Each slot in the timer wheel is a head pointer to the list of entries that must be processed during that timer tick.

Fields
time: CachePadded<UnsafeCell<Time>>

Only accessed from Registration.

inner: Option<Weak<Inner>>

Timer internals. Using a weak pointer allows the timer to shutdown without all Delay instances having completed.

When None, the entry has not yet been linked with a timer instance.

state: AtomicU64

Tracks the entry state. This value contains the following information:

The deadline at which the entry must be “fired”.

A flag indicating if the entry has already been fired.

Whether or not the entry transitioned to the error state.

When an Entry is created, state is initialized to the instant at which the entry must be fired. When a timer is reset to a different instant, this value is changed.

task: AtomicTask

Task to notify once the deadline is reached.

queued: AtomicBool

True when the entry is queued in the “process” stack. This value is set before pushing the value and unset after popping the value.

TODO: This could possibly be rolled up into state.

next_atomic: UnsafeCell<*mut Entry>

Next entry in the “process” linked list.

Access to this field is coordinated by the queued flag.

Represents a strong Arc ref.

when: UnsafeCell<Option<u64>>

When the entry expires, relative to the start of the timer (Inner::start). This is only used by the timer.

A Delay instance can be reset to a different deadline by the thread that owns the Delay instance. In this case, the timer thread will not immediately know that this has happened. The timer thread must know the last deadline that it saw as it uses this value to locate the entry in its wheel.

Once the timer thread observes that the instant has changed, it updates the wheel and sets this value. The idea is that this value eventually converges to the value of state as the timer thread makes updates.

next_stack: UnsafeCell<Option<Arc<Entry>>>

Next entry in the State’s linked list.

This is only accessed by the timer

prev_stack: UnsafeCell<*const Entry>

Previous entry in the State’s linked list.

This is only accessed by the timer and is used to unlink a canceled entry.

This is a weak reference.

Implementations

impl Entry

pub fn new(deadline: Instant, duration: Duration) -> Entry

pub fn time_ref(&self) -> &Time

Only called by Registration

pub fn time_mut(&self) -> &mut Time

Only called by Registration

pub fn is_registered(&self) -> bool

Returns true if the Entry is currently associated with a timer instance.

pub fn register(me: &mut Arc<Self>)

Only called by Registration

pub fn register_with(me: &mut Arc<Self>, handle: HandlePriv)

Only called by Registration

fn transition_to_error(&mut self)

pub fn when_internal(&self) -> Option<u64>

The current entry state as known by the timer. This is not the value of state, but lets the timer know how to converge its state to state.

pub fn set_when_internal(&self, when: Option<u64>)

pub fn load_state(&self) -> Option<u64>

Called by Timer to load the current value of state for processing

pub fn is_elapsed(&self) -> bool

pub fn fire(&self, when: u64)

pub fn error(&self)

pub fn cancel(entry: &Arc<Entry>)

pub fn poll_elapsed(&self) -> Poll<(), Error>

pub fn reset(entry: &mut Arc<Entry>)

Only called by Registration

fn upgrade_inner(&self) -> Option<Arc<Inner>>