Module tokio_timer::wheel

Modules

level

stack

Structs

Poll
Poll expirations from the wheel

Wheel
Timing wheel implementation.

Enums

InsertError

Constants

MAX_DURATION
The maximum duration of a delay

NUM_LEVELS
Number of levels. Each level has 64 slots. By using 6 levels with 64 slots each, the timer is able to track time up to 2 years into the future with a precision of 1 millisecond.

Functions

level_for