Struct tokio_timer::wheel::Wheel

pub(crate) struct Wheel<T> {
    elapsed: u64,
    levels: Vec<Level<T>>,
}

Timing wheel implementation.

This type provides the hashed timing wheel implementation that backs Timer and DelayQueue.

The structure is generic over T: Stack. This allows handling timeout data being stored on the heap or in a slab. In order to support the latter case, the slab must be passed into each function allowing the implementation to lookup timer entries.

See Timer documentation for some implementation notes.

Fields
elapsed: u64

The number of milliseconds elapsed since the wheel started.

levels: Vec<Level<T>>

Timer wheel.

Levels:

1 ms slots / 64 ms range

64 ms slots / ~ 4 sec range

~ 4 sec slots / ~ 4 min range

~ 4 min slots / ~ 4 hr range

~ 4 hr slots / ~ 12 day range

~ 12 day slots / ~ 2 yr range

Implementations

impl<T> Wheel<T> where
T: Stack,

pub fn new() -> Wheel<T>

Create a new timing wheel

pub fn elapsed(&self) -> u64

Return the number of milliseconds that have elapsed since the timing wheel’s creation.

pub fn insert(
&mut self,
when: u64,
item: T::Owned,
store: &mut T::Store
) -> Result<(), (T::Owned, InsertError)>

Insert an entry into the timing wheel.

Arguments

when: is the instant at which the entry should be fired. It is represented as the number of milliseconds since the creation of the timing wheel.

item: The item to insert into the wheel.

store: The slab or () when using heap storage.

Return

Returns Ok when the item is successfully inserted, Err otherwise.

Err(Elapsed) indicates that when represents an instant that has already passed. In this case, the caller should fire the timeout immediately.

Err(Invalid) indicates an invalid when argument as been supplied.

pub fn remove(&mut self, item: &T::Borrowed, store: &mut T::Store)

Remove item from thee timing wheel.

pub fn poll_at(&self) -> Option<u64>

Instant at which to poll

pub fn poll(
&mut self,
poll: &mut Poll,
store: &mut T::Store
) -> Option<T::Owned>

fn next_expiration(&self) -> Option<Expiration>

Returns the instant at which the next timeout expires.

pub fn poll_expiration(
&mut self,
expiration: &Expiration,
store: &mut T::Store
) -> Option<T::Owned>

fn set_elapsed(&mut self, when: u64)

fn pop_entry(
&mut self,
expiration: &Expiration,
store: &mut T::Store
) -> Option<T::Owned>

fn level_for(&self, when: u64) -> usize