Crate tokio_timer

Utilities for scheduling work to happen after a period of time.

This crate provides a number of utilities for working with periods of time:

Delay: A future that completes at a specified instant in time.

Interval A stream that yields at fixed time intervals.

Deadline: Wraps a future, requiring it to complete before a specified instant in time, erroring if the future takes too long.

These three types are backed by a Timer instance. In order for Delay, Interval, and Deadline to function, the associated Timer instance must be running on some thread.

Re-exports

pub use self::timer::Timer;

Modules

clock
A configurable source of time.

timer
Timer implementation.

Structs

Deadline
Allows a given Future to execute until the specified deadline.

DeadlineError
Error returned by Deadline future.

Delay
A future that completes at a specified instant in time.

Error
Errors encountered by the timer implementation.

Interval
A stream representing notifications at fixed interval

Functions

sleep
Create a Future that completes in duration from now.

with_default
Set the default timer for the duration of the closure.