Struct tokio_timer::Timer

pub struct Timer {
    // some fields omitted
}

A facility for scheduling timeouts

Methods

impl Timer

fn sleep(&self, duration: Duration) -> Sleep

Returns a future that completes once the given instant has been reached

fn timeout<F, E>(&self, future: F, duration: Duration) -> Timeout<F> where F: Future<Error=E>, E: From<TimeoutError>

Allow the given future to execute for at much duration time.

If the given future completes within the given time, then the Timeout future will complete with that result. If duration expires, the Timeout future completes with a TimeoutError.